pyproject.toml
[tool.poetry]
name = "langgaph-course"
version = "0.1.0"
description = ""
authors = ["Eden Marco <emarco177@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
langchain = ">=1.2"
langchain-tavily = "^0.2.0"
langgraph = ">=1.1"
python-dotenv = "^1.1.0"
black = "^25.1.0"
isort = "^6.0.1"
langchain-ollama = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
